+++ libs/test/mutate.go
@@ -1,5 +1,3 @@
-// +build unit
-
 package test
 
 import (
