+++ lite2/client_benchmark_test.go
@@ -1,3 +1,5 @@
+// +build unit
+
 package lite_test
 
 import (
